pyproject.toml
[project]
name = "tarefa-4-formula-one-races"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "nbformat>=5.10.4",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "plotly>=6.3.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]

[tool.ruff]
line-length = 180
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]
